VideoGen V2/requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
diffusers==0.27.2
transformers==4.38.2
accelerate==0.28.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
